requirements.txt
pytesseract
Pillow
beautifulsoup4
requests
langdetect
